Moved examples files from geometry_world to scene_graph
