+++ examples/scene_graph/BUILD.bazel
