+++ examples/scene_graph/bouncing_ball_vector.named_vector
