+++ examples/scene_graph/planet_rings.obj
