+++ examples/scene_graph/solar_system.h
